Django_backend/docker_test/views.py
from django.shortcuts import render
from rest_framework import viewsets, permissions

from django.contrib.auth.models import User
from .serializers import UserSerializer


class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    permission_classes = [permissions.AllowAny]
    serializer_class = UserSerializer
